Final Project/Code Releases/Release Version 2.0/thomaspay/requirements.txt
Click==7.0
Flask==1.0.2
Flask-Session==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mysql-connector==2.1.6
mysql-connector-python==8.0.13
MySQL-python==1.2.5
now==0.0.6
protobuf==3.6.1
six==1.11.0
Werkzeug==0.14.1
